Merged PR 161: Fix input when inputsim actor is recreated after a map load.
Fix input when inputsim actor is recreated after a map load.

In this case the UWorld::InitializeActorsForPlay function that handles auto-receive-input actors
may be run before the inputsim actor is created. In order to ensure it receives input events
the input must be enabled explicitly so the actor's InputComponent is pushed onto the
PlayerController's CurrentInputStack.

Also removed unnecessary labeling of the actor, because it may be omitted in non-editor builds
which can be confusing.

Related work items: #272