Merged PR 221: Fix button rest position
Fixes the button rest position so it reacts to changes in the buttons transform. This was causing issues when the button was attached to an actor with a follow component for example.

The issue stemmed from the fact that the button rest position was set to the initial position of the front face, but not updated as the button moved. The new method simply uses a local rest position and updates it with the button transform.

Related work items: #308