Merged PR 268: Use a sweep in the near interaction activation query
There were issues with thin buttons, where the input simulation fingertip would jump behind the button. This would have been fine, as there is a sweep in poke detection in case there was a collision between frames. However, the near interaction activation query was not using a sweep, and so in the second frame far interaction became active, near interaction became inactive and the sweep in poke detection was never applied.

This issue was only seen in editor, and only when scrolling the mouse wheel instead of pressing it for a poke gesture.

This change makes the near interaction activation query use a sweep to catch these cases. We should look into this in greater detail when investigation is being done for #279

Related work items: #321