Merged PR 296: Hand Constraint Component
New component to align an actor with a hand. This is specifically intended for hand menus, which need to stay in an accessible position without overlapping with the hand itself.

The component works by defining a general "zone" for placement in relation to the hand. The goal position is then computed using this zone direction by projecting onto a bounding box of the hand. Actor rotation can be configured independently.

The actor movement supports exponential smoothing like other similar movement components (follow component, manipulator). Movement can be disabled if necessary, in which case the goal position can be used in blueprints without affecting the actor transform.

Related work items: #306