Merged PR 294: Add remaining states to button
Add focused and contacted states to the PressableButtonComponent.

This goes back a bit on what I did before, the button itself no longer keeps track of what specific state it is in as this was causing everywhere the state was updated to know the details of how the states can change.
e.g. when released, should the state be `Default` or `Focused`

This version keeps track of each property individually, and deduces which state the button is in when `GetState()` is called.

Also fixed bug where the button was not updating it's internal pressed state when pressed by a far pointer.

Related work items: #285