Merged PR 305: Update pointer positions in near pointer
Update material parameter containing the poke pointer position in the near pointer.

Previously we were updating it in the hand interaction actor but, this being exclusively a near interaction concern, it is better done in the near pointer.

Also modified the default position to an unlikely value (FLT_MAX) to minimize the chance of getting highlights from disabled pointers and made sure that position is set when the pointer is disabled.

Related work items: #405