Merged PR 345: Use CalcLocalBounds in poke begin/end queries
UPrimitiveComponent::GetCollisionShape returns the primitive's world space bounds, whereas some of it's overloads, such as UBoxComponent::GetCollisionShape, return the bounds in local space. Because we were using this in poke detection, poke was behaving inconsistently for targets without a box component.

A fix is to use Search Results USceneComponent::CalcLocalBounds in poke begin/end queries as it is consistently gives local bounds.

Related work items: #486