Merged PR 460: Button icon occasionally invisible on disabled buttons
I haven't been able to repro the issue but I think the cause was setting the icon back to its initial location in `On Button Disabled`. If the button got disabled before the initial location was cached, it would cause the icon to be inside the backplate and therefore invisible.

This icon reset can be safely removed as the button will move the icon back automatically as part of `Animate Icon`

Related work items: #584