Merged PR 531: Fix bound control affordances lag
Object movement is done using the generic manipulator. If the bounds control component ticks before the manipulator, the affordances are updated to the previous frame's position causing the lag.

Change UUxtBoundsControlComponent to tick post physics to allow the affordances to properly react to external changes to the actor.

Related work items: #674