Merged PR 548: Introduce roll for vertical normals
By defining an up vector when forward is vertical, we can introduce roll to the target object. The top of the target object will face away from the user while placing on a horizontal surface.

This resolves the issue where transitioning from horizontal surfaces to non-horizontal surfaces was causing jumps in rotation. With these changes, the orientation of the target object as you transition between surfaces will be more closely aligned.

Related work items: #673