Merged PR 575: BoundsControl: Use StaticMesh assets instead of Blueprint classes for defining affordances
Now that a single `UxtGrabTargetComponent` can support multiple primitives we don't need to use separate actors for each affordance any more. Affordance instances become simply `StaticMeshComponents` on a single actor ("BoundsControlActor"), with a single `UxtGrabTargetComponent` for interaction.

As a consequence the class references for defining affordances can be replaced by simpler `StaticMesh` assets. This is a breaking change if the default affordance BPs have been replaced on a BoundsControlComponent - unlikely, but mentioned in release notes.

* Affordance class refs replaced with `StaticMesh` assets
* BoundsControlActor contains all affordances as `StaticMeshComponent`, plus a single `UxtGrabTargetComponent` for interaction.
  Reason for not using the original owning actor is to avoid problems with scale: BoundsControlActor only uses loc + rot from the transform so as to not deform affordance meshes.
* "ActiveAffordanceGrabPointers" => "GrabbedAffordances": Simply an array of grabbed affordances now. We don't need to transform the grab point any more, so can just use the existing `FUxtGrabPointerData` from the grab component.

Related work items: #706