Merged PR 647: Fix level change crash
The crash itself was caused by dereferencing a null component in `UUxtHandInteractionActor::IsNearTarget(..)` while changing level. The components passed in to this function come from an overlap query, and from some searching it looks like checking overlaps during a level change is a common source of crashes. (In this case we got an overlap containing a null component reference)

Added validation to the places that we query for overlaps.

Related work items: #901