Merged PR 709: Stabilize the grip pose from input sim to avoid movement when grabbing objects.
Grip pose is now universally used for grabbing with near or far pointers.
With input sim this leads to a regression where an object keeps rotating after
the grip action has started. On device this is not really noticeable because
grip starts when fingers are close together, but input sim starts gripping right
at the beginning of animation.

UXTools expects a stable grip pose regardless of the source (device or input sim),
so the solution here is to stabilize the grip pose: Once gripping starts the
input sim returns not simply the palm transform, but a the last palm offset from
the wrist at the grip start point, applied on top of the wrist transform.
The wrist remains stable during animation so there is no movement when grabbing.

Related work items: #1025, #1029