Fix the smoothing calculation for manipulator and pinch slider components.
These we using an incorrect frame-rate dependency that caused motion to slow down instead of speeding up with low fps.
Now the calculation matches the "1 - exp(..)" formula used by Follow, HandConstraint, and Tap-to-Place components
already.

Unfortunately this is a breaking change because the smoothing factors are interpreted differently. The release notes
include instructions for calculating equivalent values for behavior at the 60 fps target.