Merged PR 959: Fix for creeping roll offsets when using the FixedRotationToUser constraint.
The constraint has an option to exclude roll of the camera (enabled by
default). When enabled the roll of the camera is ignored (set to zero)
when applying the changed camera transform.

This also has to be taken into account when computing the fixed
transform delta initially, otherwise the initial camera roll is baked
into the delta and applied on top of the desired transform.

Related work items: #776, #852, #1018, #1103, #1104, #1108, #1421, #1427