Merged PR 1119: Fix for XRSimulation hands.
The TSoftClassPtr does not actually work for blueprint classes. Instead
a TSoftObjectPtr<UAnimBlueprint> must be used.

This also needs to be lazy-loaded outside of the constructor to avoid
running into loading order issues again (with anim BPs requiring the
PropertyAccessEditor modular feature). So the asset loading for hand
meshes and anim BPs has been moved to the BeginPlay function.

This in turn requires changes to the hand anim BP, because its BeginPlay
event is only fired when the anim BP is already in place when the
skeletal mesh begins playing. To keep the initialization in the BP
working, the InitializeAnimation event must be used instead.

Related work items: #1591